Changed to intuitive casting, removed dtype tests from basic test class

+++ heat/spatial/tests/test_distances.py
@@ -168,7 +168,7 @@ class TestDistances(TestCase):
         B = A.astype(ht.int32)
 
         d = ht.spatial.cdist(A, B, quadratic_expansion=False)
-        result = ht.array(res, dtype=ht.float64, split=0)
+        result = ht.array(res, dtype=ht.float32, split=0)
         self.assertTrue(ht.allclose(d, result, atol=1e-5))
 
         n = ht.communication.MPI_WORLD.size
@@ -185,7 +185,7 @@ class TestDistances(TestCase):
         B = A.astype(ht.int32)
 
         d = ht.spatial.cdist(A, B, quadratic_expansion=False)
-        result = ht.array(res, dtype=ht.float64, split=0)
+        result = ht.array(res, dtype=ht.float32, split=0)
         self.assertTrue(ht.allclose(d, result, atol=1e-8))
 
         B = A.astype(ht.float64)
@@ -204,7 +204,7 @@ class TestDistances(TestCase):
 
         B = A.astype(ht.int32)
         d = ht.spatial.cdist(B, quadratic_expansion=False)
-        result = ht.array(res, dtype=ht.float64, split=0)
+        result = ht.array(res, dtype=ht.float32, split=0)
         self.assertTrue(ht.allclose(d, result, atol=1e-8))
 
         B = A.astype(ht.float64)
